refactored services structure
